// client/src/app/core/lm-core.routes.ts
export enum CORE_ROUTES {
  EMPTY = '',
  HOME = 'home',
  LOGIN = 'login',
  LOGOUT = 'logout',
  SIGNUP = 'signup',
  PRICING = 'pricing',
  BILLING = 'billing',
  DASHBOARD = 'dashboard',
  TEST = 'test',
  WILDCARD = '**',
  ERROR = 'error',
  VERIFY = 'verify',
  DEDUP = 'dedup',
  SETTINGS = 'settings',
  CONTACT = 'contact',
  MEDIALAB = 'medialab',
  FEEDBACK = 'feedback'
}